Simplify delete_bag

+++ lib/moab/bagger.rb
@@ -67,14 +67,7 @@ module Moab
     # @return [NilClass] Delete the bagit files
     def delete_bag
       # make sure this looks like a bag before deleting
-      if bag_pathname.join('bagit.txt').exist?
-        if bag_pathname.join('data').exist?
-          bag_pathname.rmtree
-        else
-          bag_pathname.children.each(&:delete)
-          bag_pathname.rmdir
-        end
-      end
+      bag_pathname.rmtree if bag_pathname.join('bagit.txt').exist?
       nil
     end
